pyproject.toml
[tool.poetry]
name = "amex_env"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
wandb = "^0.12.16"
jupyter = "^1.0.0"
jupyterlab = "^3.4.2"
python-dotenv = "^0.20.0"
numpy = "^1.22.4"
pandas = "^1.4.2"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"
scikit-learn = "^1.1.1"
category-encoders = "^2.4.1"
catboost = "^1.0.6"
hydra-core = "^1.2.0"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
